started conversion of StoreController

+++ 5Bites/5Bites/Models/Store/Inventory/ProductModel.cs
@@ -8,7 +8,7 @@ namespace _5Bites.Models.Store_.Inventory
     public class ProductModel
     {
         public string Description { get; set; }
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public int Quantity { get; set; }
     }
 }
